+++ miranda/protocols/IcqOscarJ/i18n.h
@@ -43,7 +43,7 @@ BOOL __stdcall IsUSASCII(const char *pBuffer, int nSize);
 BOOL __stdcall IsUnicodeAscii(const WCHAR *pBuffer, int nSize);
 int  __stdcall UTF8_IsValid(const char *pszInput);
 
-size_t __stdcall get_utf8_size(const WCHAR *unicode);
+int __stdcall get_utf8_size(const WCHAR *unicode);
 
 char* __stdcall detect_decode_utf8(const char *from);
 
@@ -58,27 +58,27 @@ char*  __stdcall ansi_to_utf8_codepage(const char *ansi, WORD wCp);
 
 WCHAR* __stdcall ansi_to_unicode(const char *ansi);
 char* __stdcall unicode_to_ansi(const WCHAR *unicode);
-char* __stdcall unicode_to_ansi_static(const WCHAR *unicode, char *ansi, size_t ansi_size);
+char* __stdcall unicode_to_ansi_static(const WCHAR *unicode, char *ansi, int ansi_size);
 
 int   __stdcall utf8_encode(const char *from, char **to);
 int   __stdcall utf8_decode(const char *from, char **to);
 int   __stdcall utf8_decode_codepage(const char *from, char **to, WORD wCp);
-int   __stdcall utf8_decode_static(const char *from, char *to, size_t to_size);
+int   __stdcall utf8_decode_static(const char *from, char *to, int to_size);
 
 #ifdef _UNICODE
 	#define tchar_to_utf8 make_utf8_string
 	#define utf8_to_tchar_static make_unicode_string_static
-  #define utf8_to_tchar make_unicode_string
-  #define ansi_to_tchar ansi_to_unicode
-  #define tchar_to_ansi unicode_to_ansi
+	#define utf8_to_tchar make_unicode_string
+	#define ansi_to_tchar ansi_to_unicode
+	#define tchar_to_ansi unicode_to_ansi
 #else
-   __inline char* utf8_decode_func(const char *utf8) { char *ansi = NULL; utf8_decode(utf8, &ansi); return ansi; };
+	__inline char* utf8_decode_func(const char *utf8) { char *ansi = NULL; utf8_decode(utf8, &ansi); return ansi; };
 
 	#define tchar_to_utf8 ansi_to_utf8
 	#define utf8_to_tchar_static utf8_decode_static
-  #define utf8_to_tchar utf8_decode_func
-  #define ansi_to_tchar null_strdup
-  #define tchar_to_ansi null_strdup
+	#define utf8_to_tchar utf8_decode_func
+	#define ansi_to_tchar null_strdup
+	#define tchar_to_ansi null_strdup
 #endif
 
 void InitI18N(void);
